Additional tweaks and cleanup, plus a command for testing SMPP messages
