+++ kolibri_instant_schools_plugin/auth/phone_hashing_settings.py
@@ -1,4 +1,6 @@
-from ..instant_schools_settings import *
+import sys
 
-print("WARNING: `kolibri_instant_schools_plugin.auth.phone_hashing_settings` are no longer used; \
-       please use `kolibri_instant_schools_plugin.instant_schools_settings` instead.")
+print("ERROR: `kolibri_instant_schools_plugin.auth.phone_hashing_settings` are no longer used; " +
+      "please use `kolibri_instant_schools_plugin.instant_schools_settings` instead.")
+
+sys.exit(1)
